src/movie_generate_svc/requirements.txt
fastapi
uvicorn
gunicorn
jinja2
python-multipart
openai 
python-dotenv  
requests
fastapi-logger
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-openai
opentelemetry-api >= 1.10.0
opentelemetry-sdk >= 1.10.0
azure-monitor-opentelemetry 
redis
azure-ai-inference
